refactor(schema): extract public event definitions
