+++ packages/server/src/groups/event.ts
@@ -1,5 +1,7 @@
 import { EventV2 } from "@opencode-ai/core/event"
+import { PublicEventManifest } from "@opencode-ai/core/public-event-manifest"
 import { Location } from "@opencode-ai/core/location"
+import type { Definition } from "@opencode-ai/schema/event"
 import { Schema } from "effect"
 import { HttpApiEndpoint, HttpApiGroup, OpenApi } from "effect/unstable/httpapi"
 
@@ -10,33 +12,40 @@ const fields = {
   location: Schema.optional(Location.Ref),
 }
 
-const Event = Schema.Union([
-  ...EventV2.definitions().map((definition) =>
-    Schema.Struct({
-      ...fields,
-      type: Schema.Literal(definition.type),
-      data: definition.data as Schema.Struct<{}>,
-    }).annotate({ identifier: `V2Event.${definition.type}` }),
-  ),
-  Schema.Struct({
-    ...fields,
-    type: Schema.Literal("server.connected"),
-    data: Schema.Struct({}),
-  }).annotate({ identifier: "V2Event.server.connected" }),
-]).annotate({ identifier: "V2Event" })
-
-export const EventGroup = HttpApiGroup.make("server.event")
-  .add(
-    HttpApiEndpoint.get("event.subscribe", "/api/event", {
-      success: Event,
-    }).annotateMerge(
-      OpenApi.annotations({
-        identifier: "v2.event.subscribe",
-        summary: "Subscribe to events",
-        description: "Subscribe to native event payloads for the server.",
-      }),
+const schema = (definitions: ReadonlyArray<Definition>) =>
+  Schema.Union([
+    ...definitions.map((definition) =>
+      Schema.Struct({
+        ...fields,
+        type: Schema.Literal(definition.type),
+        data: definition.data,
+      }).annotate({ identifier: `V2Event.${definition.type}` }),
     ),
-  )
-  .annotateMerge(OpenApi.annotations({ title: "events", description: "Experimental event stream route." }))
+    ...(definitions.some((definition) => definition.type === "server.connected")
+      ? []
+      : [
+          Schema.Struct({
+            ...fields,
+            type: Schema.Literal("server.connected"),
+            data: Schema.Struct({}),
+          }).annotate({ identifier: "V2Event.server.connected" }),
+        ]),
+  ]).annotate({ identifier: "V2Event" })
 
-export type Event = typeof Event.Type
+export const makeEventGroup = (definitions: ReadonlyArray<Definition>) =>
+  HttpApiGroup.make("server.event")
+    .add(
+      HttpApiEndpoint.get("event.subscribe", "/api/event", {
+        success: schema(definitions),
+      }).annotateMerge(
+        OpenApi.annotations({
+          identifier: "v2.event.subscribe",
+          summary: "Subscribe to events",
+          description: "Subscribe to native event payloads for the server.",
+        }),
+      ),
+    )
+    .annotateMerge(OpenApi.annotations({ title: "events", description: "Experimental event stream route." }))
+
+export const EventGroup = makeEventGroup(PublicEventManifest.Latest.values().toArray())
+export type Event = Schema.Schema.Type<ReturnType<typeof schema>>
